\name{marijuana}

\docType{data}

\alias{marijuana}

\title{Changes in heart rate after marijuana use}

\description{
This dataset comes from a pilot study to investigate the effects of
marijuana.  Nine male subjects received three treatments each
(placebo, low dose and high dose) and the order of treatments was
balanced in a replicated 3 x 3 Latin square. Under each treatment, the
change in heart rate (beats per minute above baseline) was recorded 15
minutes after smoking and 90 minutes after smoking, producing six
measures per subject, but five of the 54 data values are
missing. These data were analyzed by Schafer (1997, Chap. 5).
}

\usage{data(marijuana)}

\format{a data frame with 9 rows and 6 variables.}

\source{
Weil, A.T., Zinberg, N.E. and Nelson, J.M. (1968) Clinical
and psychological effects of marihuana in man. \emph{Science},
62, 1234-1242.
}

\references{
Schafer, J.L. (1997) \emph{Analysis of Incomplete Multivariate
Data}. London: Chapman & Hall/CRC Press.
\cr

For example analyses of this dataset using functions in
the \code{norm2} package, see the manual \emph{NORM Package for R, Version 2}
in the library subdirectory \code{doc}.
}

\keyword{ datasets }
